requirements-gcs.txt
flask==2.3.3
flask-cors==4.0.0
numpy==1.24.3
tensorflow==2.13.0
scikit-learn==1.3.0
google-cloud-storage==2.10.0
gunicorn==21.2.0
